requirements.txt
# Core
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
httpx>=0.27.0
sse-starlette>=2.0.0
websockets>=12.0

# Google Cloud
google-cloud-aiplatform>=1.45.0
google-cloud-storage>=2.15.0
google-cloud-logging>=3.9.0
google-cloud-functions>=1.16.0
google-cloud-secret-manager>=2.20.0
google-cloud-pubsub>=2.23.0

# Document processing
unstructured>=0.12.4
unstructured-inference>=0.7.28
pdf2image>=1.17.0
pytesseract>=0.3.10
python-magic>=0.4.27
apache-tika>=2.8.0
onnxruntime>=1.17.1

# Vector DB and search
pinecone-client>=3.0.2
elasticsearch>=8.12.0
opensearch-py>=2.4.2

# Observability
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-gcp-trace>=1.5.0
opentelemetry-instrumentation-fastapi>=0.45b0
prometheus-client>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
locust>=2.24.0
mock>=5.1.0

# Dev tools
ruff>=0.2.2
mypy>=1.8.0
black>=24.1.1
isort>=5.13.2
